@@ -46,82 +46,89 @@
     "@graphql-codegen/typescript-resolvers": "^1.19.0",
     "@graphql-codegen/typescript-urql": "^2.0.5",
     "@next/bundle-analyzer": "^10.0.9",
-    "@nrwl/cli": "12.5.1",
-    "@nrwl/cypress": "12.5.1",
-    "@nrwl/eslint-plugin-nx": "12.5.1",
-    "@nrwl/jest": "12.5.1",
-    "@nrwl/linter": "12.5.1",
-    "@nrwl/next": "12.5.1",
-    "@nrwl/react": "12.5.1",
-    "@nrwl/storybook": "12.5.1",
-    "@nrwl/tao": "12.5.1",
-    "@nrwl/web": "12.5.1",
-    "@nrwl/workspace": "12.5.1",
+    "@nrwl/cli": "15.9.2",
+    "@nrwl/cypress": "15.9.2",
+    "@nrwl/eslint-plugin-nx": "15.9.2",
+    "@nrwl/jest": "15.9.2",
+    "@nrwl/js": "15.9.2",
+    "@nrwl/linter": "15.9.2",
+    "@nrwl/next": "15.9.2",
+    "@nrwl/nx-cloud": "15.3.5",
+    "@nrwl/react": "15.9.2",
+    "@nrwl/storybook": "15.9.2",
+    "@nrwl/web": "15.9.2",
+    "@nrwl/workspace": "15.9.2",
     "@storybook/addon-actions": "^6.3.0",
-    "@storybook/addon-essentials": "~6.3.0",
+    "@storybook/addon-essentials": "6.4.22",
     "@storybook/addon-knobs": "~6.3.0",
     "@storybook/addon-postcss": "^2.0.0",
-    "@storybook/react": "~6.3.0",
+    "@storybook/react": "6.4.22",
     "@svgr/webpack": "^5.4.0",
     "@testing-library/cypress": "^7.0.4",
-    "@testing-library/react": "11.2.5",
+    "@testing-library/react": "13.4.0",
     "@types/connect-redis": "^0.0.16",
     "@types/cors": "^2.8.10",
-    "@types/express": "^4.17.11",
+    "@types/express": "4.17.14",
     "@types/express-session": "^1.17.3",
     "@types/faker": "^5.1.7",
     "@types/graphql": "^14.5.0",
-    "@types/jest": "^26.0.22",
+    "@types/jest": "29.4.4",
     "@types/just-debounce-it": "^1.1.0",
     "@types/lodash": "^4.14.168",
-    "@types/react": "17.0.3",
-    "@types/react-dom": "17.0.3",
+    "@types/react": "18.0.25",
+    "@types/react-dom": "18.0.9",
     "@types/testing-library__cypress": "^5.0.8",
     "@types/webpack": "4.41.21",
     "@types/yup": "^0.29.11",
-    "@typescript-eslint/eslint-plugin": "4.3.0",
-    "@typescript-eslint/parser": "4.3.0",
+    "@typescript-eslint/eslint-plugin": "5.58.0",
+    "@typescript-eslint/parser": "5.58.0",
     "@urql/devtools": "^2.0.3",
     "@zeit/next-sass": "^1.0.1",
-    "babel-jest": "26.2.2",
+    "babel-jest": "29.4.3",
     "babel-loader": "8.1.0",
     "class-validator": "^0.13.1",
     "css-loader": "^5.2.6",
     "cypress": "^6.8.0",
     "cypress-localstorage-commands": "^1.4.5",
     "cypress-plugin-tab": "^1.0.5",
-    "dotenv": "8.2.0",
-    "eslint": "7.10.0",
+    "dotenv": "10.0.0",
+    "eslint": "8.15.0",
+    "eslint-config-next": "11.0.1",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
-    "eslint-plugin-import": "2.22.1",
-    "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-react": "7.23.1",
-    "eslint-plugin-react-hooks": "^4.2.1-alpha-aecb3b6d1",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-jsx-a11y": "6.6.1",
+    "eslint-plugin-react": "7.31.11",
+    "eslint-plugin-react-hooks": "4.6.0",
     "faker": "^5.5.1",
     "http-server": "^0.12.3",
     "husky": "^5.2.0",
-    "jest": "27.0.3",
+    "jest": "29.4.3",
+    "jest-environment-jsdom": "28.1.3",
     "newman": "^5.2.3",
-    "node-sass": "4.14.1",
+    "nx": "15.9.2",
     "pinst": "^2.1.6",
-    "prettier": "2.3.2",
-    "sass": "^1.35.1",
+    "prettier": "2.8.7",
+    "react-test-renderer": "18.0.0",
+    "sass": "1.55.0",
     "sass-loader": "^10.1.0",
     "storybook-addon-next-router": "^2.0.4",
     "storybook-formik": "^2.1.4",
     "style-loader": "^2.0.0",
-    "ts-jest": "27.0.3",
-    "ts-node": "^9.1.1",
-    "typescript": "4.0.5",
+    "ts-jest": "29.0.5",
+    "ts-node": "10.9.1",
+    "typescript": "4.9.5",
     "url-loader": "^3.0.0"
   },
   "dependencies": {
     "@chakra-ui/react": "^1.5.0",
     "@chakra-ui/theme-tools": "^1.1.7",
-    "@emotion/react": "^11.4.0",
-    "@emotion/styled": "^11.3.0",
+    "@emotion/react": "11.10.5",
+    "@emotion/styled": "11.10.5",
     "@graphql-tools/schema": "^7.1.3",
+    "@storybook/builder-webpack5": "~6.4.5",
+    "@storybook/core-server": "6.4.22",
+    "@storybook/manager-webpack5": "~6.4.5",
     "@urql/exchange-graphcache": "^4.0.0",
     "axios": "^0.21.1",
     "clsx": "^1.1.1",
@@ -136,14 +143,14 @@
     "framer-motion": "^4.1.3",
     "graphql": "^15.5.0",
     "just-debounce-it": "^1.1.0",
-    "next": "10.2.0",
+    "next": "13.1.1",
     "next-urql": "^3.0.0",
     "node-json-db": "^1.3.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-icons": "^4.2.0",
-    "react-is": "17.0.2",
-    "react-redux": "^7.2.4",
+    "react-is": "18.2.0",
+    "react-redux": "8.0.5",
     "react-select": "^4.3.0",
     "redux": "^4.1.0",
     "redux-devtools-extension": "^2.13.9",
