@@ -1,5 +1,8 @@
-module.exports = {
+const { getJestProjects } = require('@nrwl/jest');
+
+export default {
   projects: [
+    ...getJestProjects(),
     '<rootDir>/apps/multi-cart',
     '<rootDir>/libs/react-ui',
     '<rootDir>/libs/react-shared-components',
